Avatar/requirements.txt
numpy==1.23.4
face_alignment==1.3.5
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2 
numba
resampy==0.3.1
pydub==0.25.1 
scipy==1.10.1
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml  
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.3.0
gradio
gfpgan
av
safetensors
edge-tts==6.1.10
numpy==1.23.4
pillow==10.2.0
opencv-python==4.10.0.82
imageio==2.19.3
imageio-ffmpeg==0.4.7
rembg[gpu]==2.0.51       
dlib==19.24.2
face-recognition==1.3.0
face-alignment==1.3.5
mediapipe==0.10.14
librosa==0.9.2 
numba
resampy==0.3.1
pydub==0.25.1
av
safetensors
torch==2.2.1               
torchvision==0.17.1
kornia==0.6.8
gfpgan
basicsr==1.4.2
facexlib==0.3.0
scipy==1.10.1
tqdm
yacs==0.1.8
pyyaml
joblib==1.1.0
scikit-image==0.19.3
ffmpeg-python==0.2.0
requests==2.31.0
gradio